// src/pages/bookv2/hooks/useCommonBookingData.generated.ts
export type ProductForBookingFragment = { productId: string, title?: string | null, instantConfirmation?: boolean | null, bookingRequired?: boolean | null, duration?: { label?: string | null, maxDuration?: string | null, minDuration?: string | null } | null, images?: { productImages?: Array<{ isMainImage?: boolean | null, position?: number | null, productImageId: string, title?: string | null, status?: string | null, image?: { id: string, altText?: string | null, fileName: string, top3000?: string | null, top2400?: string | null, top1600?: string | null, top1400?: string | null, top1200?: string | null, top1140?: string | null, top1024?: string | null, top800?: string | null, top768?: string | null, top500?: string | null, top480?: string | null } | null }> | null } | null, uri?: { country?: string | null, city?: string | null, url?: string | null } | null, location?: { city?: string | null, country?: string | null, region?: string | null } | null, categories?: Array<{ categoryId: string }> | null, variantProducts?: Array<{ productId: string }> | null, parentProduct?: { productId: string, uri?: { url?: string | null } | null } | null, legacy?: { updatedAt?: any | null, forceAvailabilityV1?: boolean | null, excludedBookingInfo?: string | null, bookingUrl?: string | null, openDatedBooking?: boolean | null, showMerchant?: boolean | null, hasNoIndexMetaTag?: boolean | null, hidePickupOnFrontend?: boolean | null, isDisplayGooglePay?: boolean | null, isDisplayPaypal?: boolean | null, isDisplayApplePay?: boolean | null, isDisplayAfterpay?: boolean | null, defaultDeparturePoint?: string | null, isDisplayHotelPickupWarning?: boolean | null, priceFrom?: { currencyCode: string, amount: number, convertedAmount?: { amount: number, currencyCode: string } | null } | null } | null };